Refactor so make_color is shared by EditableTable as well
