+++ psi/core/enaml/util.py
@@ -4,7 +4,9 @@ log = logging.getLogger(__name__)
 import importlib.util
 from pathlib import Path
 
+from enaml.colors import Color
 from enaml.core import import_hooks
+from enaml.qt.QtGui import QColor
 
 
 def load_enaml_module_from_file(path):
@@ -38,3 +40,22 @@ def load_manifests(objects, workbench):
             load_manifests(o, workbench)
         elif hasattr(o, 'children'):
             load_manifests(o.children, workbench)
+
+
+def make_color(color, alpha=None):
+    if isinstance(color, (tuple, list)):
+        obj = QColor(*color)
+    elif isinstance(color, str):
+        obj = QColor()
+        obj.setNamedColor(color)
+    elif isinstance(color, Color):
+        return QColor(color.red, color.green, color.blue, color.alpha)
+    else:
+        raise ValueError('Unknown color %r', color)
+    if alpha is not None:
+        obj.setAlphaF(alpha)
+        if not obj.isValid():
+            raise ValueError('Cannot create QColor from %r', color)
+    return obj
+
+
